Support TSV as a "dialect" of CSV (#4)
* Open files with newline=""

The csv module documentation says this is necessary in some cases, such
as for parsing embedded newlines in quoted fields.  Refer to the
<https://docs.python.org/3/library/csv.html#id3>.

* Support TSV as a "dialect" of CSV

For seekable streams, the delimiter is sniffed from the first 1MB of
data.  This should provide enough rows to the sniffer even for datasets
with very long rows without blowing up memory usage much.

A csv.Dialect may also be specified directly to load_csv() for
programmatic usage.

* Add a --format option to support explicit dialect selection

Useful when you want to disable sniffing or when one or both of the
files aren't seekable, so sniffing doesn't work.

Thanks, @tsibley